+++ src/test/java/io/github/astrapi69/gen/tree/BaseTreeNodeTestTree.java
@@ -24,11 +24,11 @@
  */
 package io.github.astrapi69.gen.tree;
 
-import io.github.astrapi69.id.generate.LongIdGenerator;
 import lombok.Data;
+import io.github.astrapi69.id.generate.LongIdGenerator;
 
 /**
- * The  class {@link BaseTreeNodeTestTree} holds a tree for unit testing
+ * The class {@link BaseTreeNodeTestTree} holds a tree for unit testing
  */
 @Data
 public class BaseTreeNodeTestTree
@@ -49,7 +49,8 @@ public class BaseTreeNodeTestTree
 	String fifthGrandChildValue;
 	LongIdGenerator idGenerator;
 
-	public BaseTreeNodeTestTree(){
+	public BaseTreeNodeTestTree()
+	{
 		initialize();
 	}