+++ src/test/java/io/github/astrapi69/gen/tree/visitor/MergeTreeNodesVisitorTest.java
@@ -1,32 +1,67 @@
+/**
+ * The MIT License
+ *
+ * Copyright (C) 2015 Asterios Raptis
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining
+ * a copy of this software and associated documentation files (the
+ * "Software"), to deal in the Software without restriction, including
+ * without limitation the rights to use, copy, modify, merge, publish,
+ * distribute, sublicense, and/or sell copies of the Software, and to
+ * permit persons to whom the Software is furnished to do so, subject to
+ * the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be
+ * included in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+ * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+ * LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+ * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+ * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
 package io.github.astrapi69.gen.tree.visitor;
 
-import io.github.astrapi69.gen.tree.BaseTreeNode;
-import io.github.astrapi69.gen.tree.BaseTreeNodeTestTree;
-import io.github.astrapi69.id.generate.LongIdGenerator;
+import io.github.astrapi69.gen.tree.handler.ITreeNodeHandlerExtensions;
+import org.testng.annotations.Test;
 
+import io.github.astrapi69.gen.tree.BaseTreeNode;
+import io.github.astrapi69.gen.tree.BaseTreeNodeTestData;
+
+import java.util.Collection;
+
+import static org.testng.AssertJUnit.assertEquals;
+
+/**
+ * The unit test class for the class {@link MergeTreeNodesVisitor}
+ */
 public class MergeTreeNodesVisitorTest
 {
+
+	/**
+	 * Test method for {@link MergeTreeNodesVisitor#visit(BaseTreeNode)}
+	 */
+	@Test
 	public void test()
 	{
-
+		BaseTreeNode<String, Long> baseTestTree;
 		BaseTreeNode<String, Long> root;
-		BaseTreeNode<String, Long> fourthChild;
-		BaseTreeNodeTestTree testTree;
 		MergeTreeNodesVisitor<String, Long> mergeTreeNodesVisitor;
-		LongIdGenerator idGenerator;
+		Collection<BaseTreeNode<String, Long>> allTreeNodes;
 
-		idGenerator = LongIdGenerator.of(0L);
-		root = BaseTreeNode.<String, Long> builder().id(idGenerator.getNextId()).value("I'm root")
-			.build();
+		root = BaseTreeNodeTestData.getSimpleTestTree();
 
-		fourthChild = BaseTreeNode.<String, Long> builder().id(idGenerator.getNextId()).parent(root)
-			.value("I'm the fourth child").build();
+		baseTestTree = BaseTreeNodeTestData.getBaseTestTree();
+		allTreeNodes = ITreeNodeHandlerExtensions.traverse(baseTestTree);
+		assertEquals(allTreeNodes.size(), 12);
 
-		root.addChild(fourthChild);
-
-		testTree = new BaseTreeNodeTestTree();
-		mergeTreeNodesVisitor = new MergeTreeNodesVisitor<>(testTree.getRoot());
+		mergeTreeNodesVisitor = new MergeTreeNodesVisitor<>(baseTestTree);
 
 		root.accept(mergeTreeNodesVisitor);
+		allTreeNodes = ITreeNodeHandlerExtensions.traverse(baseTestTree);
+		assertEquals(allTreeNodes.size(), 15);
 	}
+
 }